+++ src/main/java/org/netbeans/modules/hyperledger/cto/parser/ErrorParserListener.java
@@ -16,27 +16,28 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.modules.hyperledger.cto.grammar;
+package org.netbeans.modules.hyperledger.cto.parser;
+
+import java.util.ArrayList;
+import java.util.List;
+import org.antlr.v4.runtime.BaseErrorListener;
+import org.antlr.v4.runtime.RecognitionException;
+import org.antlr.v4.runtime.Recognizer;
 
 /**
  *
  * @author mario.schroeder
  */
-public final class SyntaxError {
+public class ErrorParserListener extends BaseErrorListener{
+    
+    private final List<SyntaxError> syntaxErrors = new ArrayList<>();
 
-    private final String message;
-    private final int line;
-
-    public SyntaxError(String message, int line) {
-        this.message = message;
-        this.line = line;
+    public List<SyntaxError> getSyntaxErrors() {
+        return syntaxErrors;
     }
 
-    public String getMessage() {
-        return message;
-    }
-
-    public int getLine() {
-        return line;
+    @Override
+    public void syntaxError(Recognizer<?, ?> recognizer, Object offendingSymbol, int line, int charPositionInLine, String msg, RecognitionException e) {
+        syntaxErrors.add(new SyntaxError(msg, line));
     }
 }
